refactor(rag): simplify eligibility and relocate ingestion API
Restructures the RAG module so the public surface lives in app/lib/rag.py and
internals have single, well-named homes.

- Remove RAG_ALLOWED_EXTENSIONS env + parse_rag_allowed_extensions; file
  eligibility now has one source of truth (the supported-extensions set)
- Rename SUPPORTED_EXTENSIONS -> SUPPORTED_EXTENSIONS_FOR_EXTRACTION
- Remove the chunk_document wrapper; callers use DocumentChunker().chunk
- Move IngestionResult into app/rag/types.py
- Move eligibility check into app/rag/extraction as the public
  check_extraction_eligibility (raises UnsupportedFileTypeError)
- Move store_and_link_chunks (public) + _create_missing_links (private) into
  app/rag/store.py
- Define ingest_document in app/lib/rag.py; drop the now-unraisable
  FileTypeNotAllowedError exception and its plugin handler
- Delete the emptied app/rag/ingestion.py; redistribute its tests
