+++ .env
@@ -67,7 +67,6 @@ RAG_MCP_URL=http://rag-mcp:7728/mcp
 RAG_CONCURRENCY=1
 RAG_MAX_FILE_SIZE_MB=50
 RAG_STORE_BATCH_SIZE=32
-RAG_ALLOWED_EXTENSIONS=pdf,txt,doc,docx
 RAG_EXTRACTION_BATCH_SIZE=0
 
 # Memory Profiling
